requirements.txt
torch>=1.13.0
torchvision>=0.14.0
diffusers>=0.21.0
transformers>=4.25.0
gradio>=3.35.0
streamlit>=1.22.0
pillow>=9.0.0
accelerate>=0.12.0
accelerate>=0.20.0
xformers>=0.0.20